Adding content files
