+++ src/applications/edu-benefits/0994/components/VetTecProgramView.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export default function VetTecProgramView({ formData }) {
+  return (
+    <div>
+      <h4>{formData.programName}</h4>
+      <p>Start: {formData.plannedStartDate}</p>
+    </div>
+  );
+}
