+++ src/applications/edu-benefits/0994/content/trainingProgramsInformation.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export const trainingDescription = (
+  <div>
+    <p>
+      Now we’ll gather information about the training program(s) you’re
+      interested in attending.
+    </p>
+    <p>You can list up to 3 programs.</p>
+  </div>
+);
